Add files via upload
Rearranged some code in RoundRobin.java in preparation of it becoming a class that can be integrated with the rest of the classes. 
Matchmaking now uses the stat tracker array to check if any teams in the idle team linked list have yet to play each other. If so, the teams are pulled from the list and used to form a match. 
Ran into an issue where taking teams off the idle team list results in some of the remaining teams not being put into matches (when they should be). Should be an easy fix for this.